+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Select Acquisition method from controlled vocabulary list. Refs UIOR-852.
 * Order template not populating Holdings correctly.Refs UIOR-854.
 * Rename collection field name for Acquisition method. Refs UIOR-872.
+* Move "show all hidden fields" for PO or POL to Actions menu. Refs UIOR-871.
 
 ## [3.0.2](https://github.com/folio-org/ui-orders/tree/v3.0.2) (2021-11-05)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v3.0.1...v3.0.2)